[WINESYNC] dbghelp: Use local stab_nlist declaration in macho_module.c.

wine commit id 061146c2ab530a54f8323331e88c5771dd7bba49 by Jacek Caban <jacek@codeweavers.com>
