+++ dll/win32/dbghelp/image_private.h
@@ -136,6 +136,15 @@ struct image_section_map
     LONG_PTR                    sidx;
 };
 
+struct stab_nlist
+{
+    unsigned            n_strx;
+    unsigned char       n_type;
+    char                n_other;
+    short               n_desc;
+    unsigned            n_value;
+};
+
 BOOL image_check_alternate(struct image_file_map* fmap, const struct module* module) DECLSPEC_HIDDEN;
 
 BOOL elf_map_handle(HANDLE handle, struct image_file_map* fmap) DECLSPEC_HIDDEN;
