+++ sdk/tools/winesync/dbghelp.cfg
@@ -4,4 +4,4 @@ files:
   include/dbghelp.h: sdk/include/psdk/dbghelp.h
   include/wine/mscvpdb.h: sdk/include/reactos/wine/mscvpdb.h
 tags:
-  wine: 47f1c5a993ed34e5f1e4783360019613d0f9682a
+  wine: 061146c2ab530a54f8323331e88c5771dd7bba49
